requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
dataclasses-json==0.6.7
decorator==5.2.1
Deprecated==1.2.18
dune_client==1.7.10
executing==2.2.0
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jsonpickle==4.1.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
multidict==6.6.3
mypy_extensions==1.1.0
narwhals==2.0.1
ndjson==0.3.1
networkx==3.5
numpy==2.3.2
packaging==25.0
pandas==2.3.1
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
plotly==6.2.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pycoingecko==3.2.0
pydeck==0.9.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyvis==0.3.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
streamlit==1.47.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
traitlets==5.14.3
types-Deprecated==1.2.15.20250304
types-python-dateutil==2.9.0.20250708
types-PyYAML==6.0.12.20250516
types-requests==2.32.4.20250611
types-setuptools==80.9.0.20250529
typing-inspect==0.9.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
wrapt==1.17.2
yarl==1.20.1
